Comments, delt with default file names, fixed kpoints that must be optional
Here I had to make a choice regarding the default file names.
Since the name is now part of metadata of input spec, it could
be modified by users. No-one is going to do that on regular basis,
however it might be usefull in particular situations or maybe
for other plugins.
This is just a choice, no problem to go back to the use of
file names as pure internal parameters.