Migration to aiida1.0 of bands calculations, also fix a bug in pseudo part
After a bit of efforts and work around to avoid some bugs of aiida core,
now sending and parsing bands should be fixed. I'll do few more test
in the following days.
The problem with pseudos was that local_copy_list was rewritten,
not appended, so problems with more than 1 pseudos