Restructure of tests and example folder.
From the files in the folder example I selected 14 that
covers all the important situation for the understanding of the
aiida_siesta plugin. I renamed them example_. In the future,
on top of the didactic purpose, they could be used for
"integreted" tests that requires the use of a siesta installation.
The other files previously in /example are more "unit" tests.
I will move them to a folder
tests in the main folder of aiida_siesta_plugin. They are not added to
this commit, as they are not usable, at the moment, as "unit" tests
in the sense we discussed with Sebastian and Giovanni. I'll work on
that after the release of b1.
Final part of this effort is to run this 14 example to make sure that
everithing is fine. This is the priority of the next commit!