All examples are now tested and referenced in docs
Tested all the example to make sure they run on Siesta4.0.1 and
Siesta 4.1-b3. Many examples needed big changes.
Documentation has been change accordingly to reference the example.
Lots of comment added to the scripts in order to make them more
understandable for a person that approches aiida for the first time.
Only final concern left behind is the bands parsing in the case of
spin-orbit. The file shows nspin=1 and the parser works accordingly,
but I'm not sure it is correct. I also don't find very usefull
the example_ldos.py as aiida.LDOS is a unformatted file, so not
readable.