Resolved inconsistency between spec.inp and fixed filepath
During tests I realized an inconsistency in the structure of the plugin.
aiida.xml aiida.bands MESSAGES and json.time were defined as
spec.inp, but they shouldn't be exposed to the choice
of the user, in fact siesta produces them with a fixed name.
Therefore now they are just class variables _DEFAULT_
I modified the parser accordingly to get the correct names.

Also commented one option in some examples that might confuse the
users.