Fixed bugs in SiestaBaseWorkChain and added one explicative example
In SiestaBaseWorkChain there was a call to get_plugins, a function
that is not present anymore in aiida. While fixing this problem, I
realised that the entire chunk of code where get_plugins was called
is not usefull at the moment, see code for details.
Second bug concerns the wrong management of inputs of the WC.
All were defined with Required=False and then called without
checking whether they were effectivily defined.
I also added an example that I used to test the SiestaBaseWorkChain.
I thought was usefull leaving it.