First protocol commit. Proof of concept!
Introducing few ideas on how to handle protocols. They are managed
by an external function in located in the folder
aiida_siesta/workflows/functions/protocols.py. A new workchain that
make use of the protocol is introduced in
aiida_siesta/workflows/baseprotocol.py
This is just proof of concept, not usable implementation.