Creation of the class SiestaRelaxInputGenerator
Very first step towards the creation of a class that manages the
creation of inputs to be passed to the SiestaBaseWorkChain
according to a protocol hard coded inside the class (for the moment).
This class has many attributes. They will be the starting point for
the creation of a GUI. The most important attribute is "get_builder"
that returns a builder for the SiestaBaseWorkChain with filled
inputs.