A workchain that uses a protocol to relax a structure
Introduced a workchain that uses the method "get_builder" of
SiestaRelaxInputGenerator in order to have the inputs for a
SiestaBaseWorkChain that performs a relaxation.
It also return standardized outputs as decided during hackathon
in Cineca in Feb 2020.
Still need to be tested and improved. For instance discuss the
fact that a builder can't be stored into context.