Prefix in the attributs of the SiestaCalculation, no port overwrite
To have the prefix in the attributes make the interaction with
the STM plugin (and futur other plugins) easier.
It is not correct to redifine in the SiestaCalculation some
ports that the his subclass (CalcJob) already has. Therefore
now some metadata.options are modified to only get a new
default, not overwritten