Implemented constant-current option. Refactored the inputs.
Choice taken to change the inputs structure. Now the inputs are
single ports, not a parameter dictionary anymore. The new ports are
the "mode" (Str with only two allowed values, "constant-current"
and constant-height") and "value", which is the value of either the
height or the current.
Spin will be imlemented later.
There are now two separate example, one for constant current and
one for constant height.