Introduced spin option support.
It is now possible to request through the plugin the spin option
of plstm. This feature is only implemented in the modern version
of plstm, therefore a warning is produced every time the user
request spin option but uses an old version of plstm.
The check on the plstm version is based on the name of the output.
Something more robust should be implemented in the future.