Restructured the STM WorkChain and few other changes.
Compleatly changed the STM WorkChain. Now it does not presents protocols
and all the spin options has been included.
Because in the LDOS block the energy is not defined respect to the Fermi
energy, I decided to separate the relax/scf part, from the generation of
the .LDOS. So the workflows run a separate calculation to the .LDOS once
the Fermi energy is known by the previous siesta calculation.
For this, I had to introduce the restart feature in SiestaBaseWorkChain.
The stm_array output is different for the non collinear case, as STM
calculation in all the spin components are performed.
Actually the syntax of stm_array is modified in the plugin as well.
The get_stm_image.py file is modified accordingly. The
tests are modified accordingly.