Little revamping of siesta parser, fixes also a corner case bug.
The siesta parser has been simplified in order to make it more
readable and facilitate future improvements in the error management.
Also a small bug has been corrected. In the previous parser version,
the presence of the bands_file was checked before the MESSAGES file
analysis. Therefore in the situation of not scf convergence, the
process was excepted because of missing bands_file. However this is not
desired, we want the parser to check first if there is convergence.
If the calculation is converged, then we check if the bands_file is
present.