Improved error handling in parser. Added all the necessary tests.
The siesta parser has been improved in the effort to make more clear and
complete the treatment of some common failures of siesta. The
keyword GeometryMustConverge is now always added to the siesta run
because it makes mach more simple the treatment of error when
the geometry doesn't converge.
Few tests have been added to cover the new error handling part.