Adapted the psf and psml data classes in order to support the new Group system.
From aiida 1.2.0 the class `Group` have been modified to implement new features.
This has implications in our plugin because the PseudoFamilies are Groups.
In particular there is now an explicit implementation of PsmlFamily and PsfFamily
as subclasses of the `Group` class. Moreover PsmlFamily and PsfFamily are
registred as entry points making possible to reload the correct
sub class when reloading from the database.
For more info see: https://github.com/aiidateam/aiida-core/pull/3882