Introduction of the bandgap workchain.
Created as a subclass of SiestaBaseWorkChain. Also a little change in the
handling of herror has been added. The problems concerning the bands
parsing has been extended to treat the case when an incomplete bands
file is retrieved. This error and the complete absence of bands file are
now recognized by SiestaBaseWorkChain and a specific error is returned.
Missing the documentation.