Introduced possibility for custom protocols, moved the heuristics treatement.
The treatement of the atom_heuristics has been moved in the class ProtocolRegistry
and implemented in the methods _get_param and _get_basis.
Extended as well the cases that the atom_heuristics support.
Moreover added the possibility to define an environment variable AIIDA_SIESTA_PROTOCOLS
in which a user can indicate a .yaml file where he specify his/her own protocol.
The entries `pseudo_family`, `parameters` and `basis` are mandatory and their presence is
checked in the __init__.
Missing documentation and maybe adding the spin options?