Structure of tests completed. Also added a chech in generator_metaclass.py
A cheched is added to InputsGenerator metaclass in generator_metaclass.py
in order to make sure that the attribute _workchain_class defined in sublasses
has an attribute `get_builder()`. This ensure we are passing a process and
not a random variable.
The structure of the tests is now in place, but much more needs to be
added to have a full set of tests.