First set of tests for the workchains.
Cover the error handler of the SiestaBaseWorkChain and the setups/checks
of all the other WorkChains. Missing the tests on the postprocesses.
Also fixed a small bug in the SiestaBaseWorkChain. I decided to not expose
the "retrieved" output of a SiestaCalculation as output of the WorkChain,
therefore I can not request it in outputs in case of bands failure.