Update STM tests, improved a bit the STM WC.
Writing STM tests I found out that I implemented in a useless way
the checks on the good termination of the stm calculation in the
STM WorkChain, now this is improved.

Also there was an error in the generate_wc_node, now fixed.

Still missing to test the non-collinear part of the STMWorkChain