Fix requirement of exposed ports.
The feature to impose the exposed input ports to be "not-required" was due to the
fact that users can now iterate over this ports name, defining them
in the `iterate_over` dictionary.
However, when the `iterate_over` port is not accepted (new feature implemented
recently), it has no sense to impose that the exposed inputs are not-required.
This commit fixes the problem.