Introduced the input generator for SiestaCalculation.
The protocol system can be used for SiestaCalculations
as well, not only for WorkChains.
Only changes are that metadata now host options and that
there is no port for psedo_family. The port of
pseudos must be used. The protocol therfore is now implementing
the way to get pseudos from a family name.
I left this design also for all the WorkChains. It gives more
flexibility to manually change one single pseudo.