Changed the handle of kpoints in BandsData
In the outputs bands (BandsData), now the kpoints are set
using the output structure cell (or input one if no relaxation).
Even if the cell was set in the `bandskpoints`, it is overidden.
Anyway a check is added to impose that the cell in `bandskpoints`
is the same as the one of the input structure.
In rare cases when this imposition is limiting, the user can now set
kpoints without cell.