Rewamp of BandGapWorkChain.
Added the following feature to the BandGapWorkChain:
if no `bandkpoints` is requested in input, the WorkChain
applies some logic to run cleverly the bands calculation.
In particular it is detected if the run is a relaxation (1) or
a normal single point calculation (2).
If (1), the relaxation is run and an extra step at the end is
added to separatly get the bands after constructing the kpoints
path automatically using the output cell (that might be very
different from the input one).
If (2), the bandskpoints input is added straight away.