Use "if .. is not None" and small inpu_gen change
It is better to always do explicitly "if .. is not None" instead
of "if ,,,". In fact the last one is False not only for None
but also for empty dictionaries and lists.
Also changed the get_inputs_dict() of inputs_generators in order
to return only the defined values, not also the None values.
In this way it can be passed directly for submission.