Fixed bugs of `SequentialConverger`.
The parsing of `converged_target_value` output of a convergence workchain is now
done only when the convergence has been successful.
The output `converged_target_value` is returned only if the attribute
`last_target_value` is present in `self.ctx`. It might not be present
if all the convergence workchains fail.