Completly change direction: now Ion file is Data
The IonData type has been implemented to host ion files.
From there the orbitals are retrived on the fly.
This means that the file is the only stored object and not
a dedicated data type for orbitals.
Ion files are now parsed by the parser and punt in an output namespace.
The IonData also return the PaoModifier and
the `get_pao_block`. This part will be improved.