Added support for pdos parsing.
The PdosData class takes care of saving the .PDOS file in the database.
The idea is to have methods of this class that make use of sisl to return
the info in the .PDOS file in a pythonic way. For the moment only returnig
arrays of energy and pdoses separately.
The retrieve of the .PDOS file and its parsing are added to the plugin.