First stage of support for floating orbitals
Floating orbitals are specified in the 'basis' input, with arbitrary
names (but also with the related atomic symbol), and atomic positions.

Internally, the input structure is augmented with the fake atoms corresponding
to the floating orbitals, but this object is never stored, only used to generate
the fdf file.

After the calculation, the parser strips away the fake atoms and fake forces from
the structure and forces.

New examples: 'example_ghost.py' and 'example_ghost_relax.py', for demonstration
purposes. Note in particular the implementation of a two-step pseudopotential assignment
that can eliminate the need to provide pseudpotential mappings for 'alternate' species.