Parsing of Ions from floating_orbs. Remove assumptions on pseudos.
Floating orbitals are caracterized by a "name", a "symbol" and a
"position". Siesta requires one pseudo for each "name". The assumption
to use a pseudo assigned to "symbol" in case one assigned to "name" is not
found might lead to problems. Therefore it is removed. If there is no
one pseudo for each "name", error is returned.
Added the parsing of the IonData files associated to floating_orbs.