Now the IonFiles are accepted as inputs of a SiestaCalculation.
This required many adjustments, firstly to make the `pseudos` input
optional and apply all the necessary checks.
Since siesta do not allow the mixing of ions ad pseudos for the same
calculation, we impose that, if an ion is set in input, all the pseudo will
be ignored and one ion for each specie (including floating sites)
must be present.
Out of necessity, startshere the introduction of the validators. They
will be largely used in the coming refactoring.