Update examples, the Lua script, and reference data for new-neb
Some options carried over from the neb branch were outdated.

The NEB script now incorporates a helper function to make it work
with the vanilla flos library (without the material in a proposed
MR by A. Akhtar).

In the utils/neb.py, a copy of the function to parse the retrieved folder
has been added, not yet as a calcfunction.