Just a vacancy ghost in end-point relaxation. Utils versions
As expected, we get Cholesky errors due to the ill-conditioning of
the basis set when an atom and its ghost are on the same site.

To proceed from the initial stage of end-point relaxation, we
only put a ghost in the vacancy site.

The generation of the path has exposed a few incompatibilities with
the versions of some utils routines. They have been fixed. Actually,
for the vacancy exchange problem it might not be needed to use the
"mid-point generation" method at all.