Restructured the optimizer.
Now the user can select the schema for optimization playing with the variables
"global_energy_shift", "global_split_norm" and "charge_confinement".
Added the possibility to run a check with an extra orbital after the basis
sizes checks. The orbital to add must be specified in the
`add_orbital` input.
Finally allow to define a siesta_base.basis input with info that will be carried
on along the entire optimization. For instance pao-non-perturbative-polarization
or reparametrize-pseudos. This input can not contain pao-basis-sizes or
pao blocks. It allows energy shift and split norm, but these will be
removed in the last optimization step.